pyproject.toml
[project]
name = "memory-agent"
version = "0.0.1"
description = "A memory agent for case management"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.34,<0.3.0",
    "langchain-openai>=0.2.1",
    "langchain-anthropic>=0.2.1",
    "langchain>=0.3.1",
    "langchain-core>=0.3.8",
    "python-dotenv>=1.0.1",
    "langgraph-sdk>=0.1.32",
    "trustcall>=0.0.26",
    "firebase-admin>=6.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-cloud-firestore>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[tool.setuptools]
package-dir = {"" = "src"}
packages = ["memory_agent"]

